chore: moved playlist menu to features/winamp/menu
