+++ src/features/winamp/menu/index.ts
@@ -1,0 +1,1 @@
+export * from "./view";
